🔨 Cancel system added.

+++ commands/example.js
@@ -22,12 +22,13 @@ export default {
         }
 
         command () {
-
+            
             this.message.delete()
 
-            let time     = this.arguments.filter(x => x.match(/\d+(\s+)?s((econd)?(s)?)/g) || x.match(/^\d+$/g)),
-                maxUsers = 1,
-                users    = []
+            let time       = this.arguments.filter(x => x.match(/\d+(\s+)?s((econd)?(s)?)/g) || x.match(/^\d+$/g)),
+                maxUsers   = 1,
+                users      = [],
+                msg_status = false
 
             const embed = new Discord.RichEmbed()
 
@@ -44,63 +45,91 @@ export default {
             embed.setTitle('BLP • 1 vs 1')
             embed.setDescription('Please react to the following message to join the matchmaking. Be careful! Message will be deleted in ' + time + ' seconds.\n' + maxUsers + ' users are needed to process matchmaking.')
 
-            this.message.channel.send(embed).then(message => message.react('✅').then(() => {
+            this.message.channel.send(embed).then(message => message.react('✅').then(() => message.react('🚫')).then(() => {
 
                 const filter = (reaction, user) => {
-                    return ['✅'].includes(reaction.emoji.name) && user.id !== this.client.user.id
+                    return ['✅', '🚫'].includes(reaction.emoji.name) && user.id !== this.client.user.id
                 }
     
                 message.awaitReactions(filter, { max: 1, time: time * 1000 }).then(collected => {
 
                     const reaction = collected.first()
+
                     if (reaction) {
-                        if (reaction.emoji.name === '✅') {
-                            const emojiArray = Array.from(reaction.users).filter(x => x[0] !== this.client.user.id)
-                            if (emojiArray) {
+                        const emojiArray = Array.from(reaction.users).filter(x => x[0] !== this.client.user.id)
+                        if (emojiArray) {
+                            if (reaction.emoji.name === '✅') {
                                 if (emojiArray.length > maxUsers) {
                                     reaction.remove()
                                 }
+                            } else if (reaction.emoji.name === '🚫') {
+                                if (emojiArray.map(x => x[0]).filter(x => x === this.message.author.id).length > 0) {
+                                    message.delete().then(() => {
+                                        msg_status = true
+                                        const embed = new Discord.RichEmbed()
+
+                                        embed.setAuthor(`Requested by ${this.message.author.username}#${this.message.author.discriminator}`, this.message.author.avatarURL)
+                                        embed.setFooter(`By ${this.client.user.username} and Ness#9999`, this.client.user.displayAvatarURL)
+                                        embed.setTitle('Match making has been cancelled!')
+                                        embed.setDescription(`${this.message.author.username} has successfully cancelled the matchmaking.`)
+
+                                        this.message.channel.send(embed).then(message => {
+                                            setTimeout(() => {
+                                                message.delete()
+                                            }, 2500)
+                                        })
+
+                                    })
+
+                                }
+
                             }
+
                         }
+
                     }
+
                 })
+                
                 setTimeout(() => {
-                    users = Array.from(message.reactions.get('✅').users).filter(x => x[0] !== this.client.user.id).map(x => x[1])
-                    if (users.length === maxUsers) {
-                        message.guild.createChannel(users.map(x => x.username).join('-'), {permissionOverwrites: users.map(x => x = {
-                            id: x.id,
-                            allow: ['VIEW_CHANNEL']
-                        }).concat(Array.from(this.message.guild.roles).map(x => x[0]).map(x => x = {
-                            id: x,
-                            deny: ['VIEW_CHANNEL']
-                        }))}).then(channel => {
+                    if (!msg_status) {
+                        users = Array.from(message.reactions.get('✅').users).filter(x => x[0] !== this.client.user.id).map(x => x[1])
+                        if (users.length === maxUsers) {
+                            message.guild.createChannel(users.map(x => x.username).join('-'), {permissionOverwrites: users.map(x => x = {
+                                id: x.id,
+                                allow: ['VIEW_CHANNEL']
+                            }).concat(Array.from(this.message.guild.roles).map(x => x[0]).map(x => x = {
+                                id: x,
+                                deny: ['VIEW_CHANNEL']
+                            }))}).then(channel => {
+
+                                const embed = new Discord.RichEmbed()
+
+                                embed.setAuthor(`Requested by ${this.message.author.username}#${this.message.author.discriminator}`, this.message.author.avatarURL)
+                                embed.setFooter(`By ${this.client.user.username} and Ness#9999`, this.client.user.displayAvatarURL)
+                                embed.setTitle(`${users.map(x => x.username).join(' and ')}'s channel`)
+                                embed.setDescription('Here you can start talking!')
+
+                                channel.send(embed)
+
+                                channel.send(users.map(x => x = `<@${x.id}>`).join('')).then(message => message.delete())
+
+                            })
+                        } else {
 
                             const embed = new Discord.RichEmbed()
 
                             embed.setAuthor(`Requested by ${this.message.author.username}#${this.message.author.discriminator}`, this.message.author.avatarURL)
                             embed.setFooter(`By ${this.client.user.username} and Ness#9999`, this.client.user.displayAvatarURL)
-                            embed.setTitle(`${users.map(x => x.username).join(' and ')}'s channel`)
-                            embed.setDescription('Here you can start talking!')
+                            embed.setTitle(`Error during channel creation:`)
+                            embed.setDescription('A group must have ' + maxUsers + ' members!')
 
-                            channel.send(embed)
+                            message.channel.send(embed).then(message => setTimeout(() => message.delete(), 5000))
 
-                            channel.send(users.map(x => x = `<@${x.id}>`).join('')).then(message => message.delete())
-
-                        })
-                    } else {
-
-                        const embed = new Discord.RichEmbed()
-
-                        embed.setAuthor(`Requested by ${this.message.author.username}#${this.message.author.discriminator}`, this.message.author.avatarURL)
-                        embed.setFooter(`By ${this.client.user.username} and Ness#9999`, this.client.user.displayAvatarURL)
-                        embed.setTitle(`Error during channel creation:`)
-                        embed.setDescription('A group must have ' + maxUsers + ' members!')
-
-                        message.channel.send(embed).then(message => setTimeout(() => message.delete(), 5000))
-
+                        }
+                        
+                        message.delete()
                     }
-                    
-                    message.delete()
                 }, time * 1000)
 
             }))
